+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- Consul service registration and de-registration
+- Signal handling (SIGINT on Windows, SIGINT + SIGTERM on Linux) for Ctrl-C cleanup
+
 ## 0.0.1
 
 - Initial version